requirements.txt
requests
pandas
textblob
sqlalchemy
python-dotenv
streamlit
plotly